requirements.txt
# Core web framework
flask==2.3.3

# Twilio integration
twilio==9.6.0

# Environment variables
python-dotenv==1.1.0

# Web server for production
gunicorn==21.2.0

# HTTP requests
requests==2.32.3

# Google Cloud and Dialogflow
google-cloud-dialogflow==2.25.0
google-api-core==2.12.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
protobuf==4.25.2
grpcio==1.60.0
grpcio-status==1.60.0

# Asynchronous HTTP client (used by Twilio)
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
frozenlist==1.6.0
multidict==6.4.3
yarl==1.20.0